requirements.txt
Django==4.2.0
djangorestframework==3.14.0
elasticsearch==8.7.0
